11_PDF_Pipeline/requirements.txt
# PDF 處理 Pipeline 依賴套件

# PDF 處理
PyMuPDF>=1.23.0

# 圖片處理
Pillow>=10.0.0

# 其他
pathlib2>=2.3.7; python_version < '3.4'
